Refactored: moved FgsCli to 'search' package.
