+++ tetrad-cli/src/main/java/edu/cmu/tetrad/cli/search/FgsCli.java
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA 02110-1301  USA
  */
-package edu.cmu.tetrad.cli;
+package edu.cmu.tetrad.cli.search;
 
 import edu.cmu.tetrad.cli.util.Args;
 import edu.cmu.tetrad.data.BigDataSetUtility;
